+++ app/src/main/java/com/example/bundlebundle/cart/CartContentFragment.kt
@@ -51,8 +51,6 @@ class CartContentFragment : Fragment() {
         fragmentManager = parentFragmentManager
         intent = requireActivity().intent
 
-
-
         cartTab = binding.tabLayout
         setTab(cartTab)
         setTabListeners()
