+++ app/src/main/res/layout/activity_order_step.xml
@@ -172,6 +172,7 @@
             android:paddingTop="2dp">
 
             <TextView
+                android:id="@+id/btn_make_order_with_address"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:paddingLeft="7dp"
